Move RequestPermissionsResult out of ActivityStreams
